--- gems/smithy/lib/smithy/anvil.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Smithy
-  module Anvil # :nodoc: all
-  end
-end
-
-require_relative 'anvil/view'
-require_relative 'anvil/client'
