--- gems/smithy/lib/smithy/anvil/client.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Smithy
-  module Anvil
-    # @api private
-    module Client; end
-  end
-end
-
-require_relative 'client/views'
-require_relative 'client/endpoint_parameter'
-require_relative 'client/operation_examples'
-require_relative 'client/request_response_example'
