--- gems/smithy/lib/smithy/anvil/client/views.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'views/client'
-require_relative 'views/customizations'
-require_relative 'views/endpoint_parameters'
-require_relative 'views/endpoint_provider'
-require_relative 'views/errors'
-require_relative 'views/gemspec'
-require_relative 'views/module'
-require_relative 'views/shapes'
-require_relative 'views/plugins/endpoint'
-require_relative 'views/rubocop'
-require_relative 'views/specs/endpoint_provider_spec'
-require_relative 'views/specs/spec_helper'
-require_relative 'views/types'
