--- gems/smithy/lib/smithy/anvil/client/views/customizations.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-module Smithy
-  module Anvil
-    module Client
-      module Views
-        # @api private
-        class Customizations < View; end
-      end
-    end
-  end
-end
