+++ gems/smithy/lib/smithy/templates/client/client.erb
