+++ gems/smithy/lib/smithy/templates/client/endpoint_parameters.erb
