+++ gems/smithy/lib/smithy/templates/client/endpoint_provider.erb
