+++ gems/smithy/lib/smithy/templates/client/errors.erb
