+++ gems/smithy/lib/smithy/templates/client/gemspec.erb
