+++ gems/smithy/lib/smithy/templates/client/module.erb
