+++ gems/smithy/lib/smithy/templates/client/plugins/endpoint.erb
