+++ gems/smithy/lib/smithy/templates/client/rubocop.erb
