+++ gems/smithy/lib/smithy/templates/client/shapes.erb
