+++ gems/smithy/lib/smithy/templates/client/specs/endpoint_provider_spec.erb
