+++ gems/smithy/lib/smithy/templates/client/specs/spec_helper.erb
