+++ gems/smithy/lib/smithy/views.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Smithy
+  # @api private
+  module Views; end
+end
+
+require_relative 'views/view'
+require_relative 'views/client'
