+++ gems/smithy/lib/smithy/views/client/customizations.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Smithy
+  module Views
+    module Client
+      # @api private
+      class Customizations < View; end
+    end
+  end
+end
